requirements.txt
# Core AI / Embedding
google-genai>=1.0.0
lancedb>=0.6.0
pyarrow>=15.0.0

# PDF extraction
PyMuPDF>=1.24.0

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=23.2.0

# HTTP / networking
httpx>=0.27.0

# Data / ML
numpy>=1.26.0
scikit-learn>=1.4.0

# Google Cloud (optional — for Vertex AI / Discovery Engine)
google-cloud-storage>=2.14.0
google-cloud-discoveryengine>=0.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
